+++ CHANGELOG.markdown
@@ -3,7 +3,7 @@
 ## Current
 
 * Migrate from v1beta to v1 of the LIFX HTTP API.
-* Deprecated `HTTPSession -setLightsPower:power:duration:completionHandler:` and `HTTPSession -setLightsColor:color:duration:powerOn:completionHandler:`. Use `HTTPSession -setState:power:color:brightness:duration:completionHandler:` instead.
+* Deprecated `HTTPSession -setLightsPower:power:duration:completionHandler:` and `HTTPSession -setLightsColor:color:duration:powerOn:completionHandler:`. Use `HTTPSession -setLightsState:power:color:brightness:duration:completionHandler:` instead.
 * Deprecated `LightTarget -setColor:brightness:power:duration:completionHandler:`. Use `LightTarget -setState:brightness:power:duration:completionHandler:` instead.
 * Add support for iOS 8.0+.
 * Add support for Swift 2/Xcode 7.0.